\name{get_cover}

\alias{get_cover}

  
\title{get cover}

\description{
get the cover
}

\usage{
get_cover(regions, sample, env=NULL, adjust=TRUE)
}

\arguments{
\item{regions}{??}
\item{sample}{??}
\item{env}{SpatRaster with environmental data or NULL}
\item{adjust}{logical. If \code{TRUE} then the cover is used?}
}

\value{
numeric 
}

\examples{
2 + 2
}

\keyword{spatial}

