\name{get_samplesize}

\alias{get_samplesize}

  
\title{get_samplesize}

\description{
get the branch length
}

\usage{
get_samplesize(range, fun, nmin=10, minsize=10000)
}

\arguments{
\item{range}{Spat?}
\item{fun}{function}
\item{nmin}{positive integer}
\item{minsize}{positive integer}
}

\value{
numeric 
}

\examples{
2 + 2
#get_samplesize( ) 
}

\keyword{spatial}

